\name{plottingFunctions-package}
\alias{plottingFunctions-package}
\alias{plottingFunctions}
\docType{package}
\title{
\packageTitle{plottingFunctions}
}
\description{
\packageDescription{plottingFunctions}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{plottingFunctions}
\packageIndices{plottingFunctions}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{plottingFunctions}

Maintainer: \packageMaintainer{plottingFunctions}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
